pony.txt
Line #207 to #252 changed.
The crystal empire now works and is limited to the northern region of North America, you made a mistake because you used region_name instead of just region also the limit on the trigger was broken. Easy way to test is to do testevent [event id] this will give you info on what the triggers are actually looking for what requirements you meet.