pony.txt
Made the crystal empire fire properly and core the entirety of north america if a pony nation has a core in the northern region. It should only fire once. Just to let you know region_name is not a command instead it is region, this meant that the game was reading it as, any/all provinces you can see how the game is reading events by using the command testevent in the console during gameplay. I changed it to read "if one of the pony countries have at least one core in the northern region, this event can fire". ROOT grabs the country that fired the event which would be one of the pony countries (this allow players to play as pegasi and the other ponies and still get the crystal event).